backend/services/8-agentic-ai-service/requirements.txt
fastapi
uvicorn[standard]
crewai==0.30.11
crewai-tools==0.1.7
langchain-groq
python-dotenv
requests
ortools